Remove the creation of individual pxe node conf files
This is the final step 3 making the ansible task "create pxe boot
data" faster. It depends on the previous 2 steps. In step 1
(https://github.com/CSC-IT-Center-for-Science/ansible-role-pxe_config/pull/17),
pxe_nodes.json containing the PXE boot data for all nodes was
created. In step 2
(https://github.com/CSC-IT-Center-for-Science/ansible-role-pxe_bootstrap/pull/4),
boot.py was changed to use that file, and now step 3 gets rid of the
individual PXE node conf files.