Install ansible in kickstart %post section
CentOS 7.4 moved ansible from the EPEL repo to the CentOS extras repo.
For some reason it doesn't work to enable extras as a repo in
kickstart, thus install it afterwards in the %post section, when the
standard CentOS repos, including extras, have been setup.

To make this work one also needs to remove ansible from the
kickstart_packages lists in group_vars.